moved function safe_json_dumps
